backend/requirements.txt
Django==4.2.2
gunicorn==20.1.0
prometheus-client==0.17.0
django-rest-framework==0.1.0

psycopg2-binary
python-json-logger
requests

ipython